requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
pydantic>=2.0.0,<3.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# ── CLI ───────────────────────────────────────────────────────────────────────
typer[all]>=0.9.0
rich>=13.7.0

# ── LLM SDKs ─────────────────────────────────────────────────────────────────
openai>=1.12.0
anthropic>=0.28.0

# ── Evaluators ────────────────────────────────────────────────────────────────
ragas>=0.1.0,<0.2.0
datasets>=2.14.0
deepeval>=0.21.0

# ── Data / Storage ────────────────────────────────────────────────────────────
# sqlite3 is part of Python stdlib
numpy>=1.26.0
pandas>=2.1.0

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit>=1.32.0
plotly>=5.18.0

# ── Reporting / GitHub ────────────────────────────────────────────────────────
PyGithub>=2.2.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.23.0
